refactor: MoveCondition 의 확장 범위 축소
움직임에 집중된 구현과 테스트를 위해 MoveCondition 에서 바로 RandomNumber를
생성하는 것으로 변경. 그에 때라 자동차의 움직임 명세는 Car에서 진행된다.
