+++ src/main/java/racingcar/domain/rule/RandomMoveCondition.java
@@ -1,31 +1,14 @@
 package racingcar.domain.rule;
 
-import racingcar.domain.util.RandomUtil;
-import racingcar.exception.MoveConditionOutOfBoundException;
+import java.util.Random;
 
-public class RandomMoveCondition implements MoveCondition {
-    private static final int MINIMUM_BOUND = 0;
-    public static final int MAXIMUM_BOUND = 9;
-    private final static int MOVABLE_BOUND = 4;
-
-    private final RandomUtil<Integer> randomUtil;
-
-    public RandomMoveCondition(RandomUtil<Integer> randomUtils) {
-        this.randomUtil = randomUtils;
-    }
+public class RandomMoveCondition implements MoveCondition<Integer> {
+    private static final Random random = new Random();
 
     @Override
-    public boolean isMovable() {
-        int randomNumber = randomUtil.generate(MAXIMUM_BOUND);
+    public Integer generate(int bound) {
+        int randomNumber = random.nextInt(bound + 1);
 
-        validateTRandomNumberBound(randomNumber);
-
-        return randomNumber >= MOVABLE_BOUND;
-    }
-
-    private void validateTRandomNumberBound(int randomNumber) {
-        if (!(MINIMUM_BOUND <= randomNumber && randomNumber <= MAXIMUM_BOUND)) {
-            throw new MoveConditionOutOfBoundException();
-        }
+        return randomNumber;
     }
 }
