--- src/main/java/racingcar/domain/util/RandomUtil.java
@@ -1,5 +1,0 @@
-package racingcar.domain.util;
-
-public interface RandomUtil<T> {
-    T generate(int bound);
-}
